+++ src/helpers/delegation/compound.ts
@@ -1,26 +1,5 @@
 import { DelegateWithPercent, DelegateWithBalance } from '@/helpers/interfaces';
-
-type QueryParams = {
-  first: number;
-  skip: number;
-  id: string;
-  orderBy?: string;
-};
-
-abstract class StandardConfig {
-  abstract getDelegatesQuery(params: QueryParams): Record<string, any>;
-  abstract formatDelegatesResponse(
-    response: Record<string, any>
-  ): DelegateWithPercent[];
-
-  abstract getDelegateQuery(id: string): Record<string, any>;
-  abstract formatDelegateResponse(
-    response: Record<string, any>
-  ): DelegateWithBalance;
-
-  abstract initializeUser(address: string): DelegateWithPercent[];
-  abstract getContractDelegateMethod(): { abi: string[]; action: string };
-}
+import { StandardConfig, QueryParams } from './standardConfig';
 
 export class CompoundGovernorConfig extends StandardConfig {
   getDelegatesQuery(params: QueryParams): Record<string, any> {
@@ -152,15 +131,5 @@ export class CompoundGovernorConfig extends StandardConfig {
       abi: ['function delegate(address delegatee)'],
       action: 'delegate'
     };
-  }
-}
-
-export function createStandardConfig(standard: string): StandardConfig {
-  switch (standard) {
-    case 'compound-governor':
-      return new CompoundGovernorConfig();
-
-    default:
-      throw new Error(`Unsupported standard: ${standard}`);
   }
 }
